Renamed index from Simple to DiskANN
